Rename directory
We no longer have to make distinction between mapgl/maplibre
map and leaflet map as we don't have leaflet dependency anymore.
