+++ src/components/map/MapFooter.tsx
