+++ src/components/map/Stop.tsx
