+++ src/routes.ts
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import { useParams } from 'react-router-dom';
 import { Main } from './components/Main';
-import { Map } from './components/mapgl-map';
+import { Map } from './components/map';
 import { RoutesAndLinesPage } from './components/RoutesAndLinesPage'; // eslint-disable-line import/no-cycle
 
 export enum Path {
